requirements.txt
Django>=4.2,<5.0
mssql-django>=1.3.0
djangorestframework
drf-yasg

django-reversion
django-cors-headers


redis
django-redis

gunicorn